Extend tests and extract functions into util clases
phovea/generator-phovea#425
